split code in components
