+++ src/components/common/SearchBar/SearchBar.js
@@ -1,66 +1,9 @@
 import React, { useState } from 'react';
-import { Input, Modal, DatePicker, Form } from 'antd';
+import { Input, Modal } from 'antd';
 import { SearchOutlined } from '@ant-design/icons';
 import SearchResults from './SearchResults';
 import 'antd/dist/antd.dark.css';
-import moment from 'moment';
-
-function InputSearch(props) {
-  const { setSearchTerm, searchTerm, searchCriteria } = props;
-  const inputTypes = { SEARCH: 'search', DATE: 'date' };
-  const inputType =
-    searchCriteria === 'dateRange' ? inputTypes.DATE : inputTypes.SEARCH;
-  const { RangePicker } = DatePicker;
-
-  const dateFormat = 'MM/DD/YYYY';
-
-  //handle search term change
-  const handleSearchTermChange = e => {
-    setSearchTerm(e.target.value);
-    console.log('handleSearchTermChange', searchTerm);
-  };
-
-  // handle date range change
-  const handleRangePickerChange = value => {
-    console.log('value', value);
-    const dateRange = value.map(date => date.format(dateFormat));
-    console.log('dateRange', dateRange);
-    const formattedDateRange = dateRange.join(' - ');
-    setSearchTerm(formattedDateRange);
-  };
-
-  if (inputType === inputTypes.SEARCH) {
-    return (
-      <div>
-        <label htmlFor="searchTerm"></label>
-        <Input
-          className={`ant-input`}
-          type={inputType}
-          id="searchTerm"
-          onChange={handleSearchTermChange}
-          value={searchTerm}
-        />
-      </div>
-    );
-  }
-
-  if (inputType === inputTypes.DATE) {
-    return (
-      <div>
-        <Form.Item label="" name="Date" initialValue="">
-          <RangePicker
-            defaultValue={[moment(), moment()]}
-            format={dateFormat}
-            placement="bottomLeft"
-            allowEmpty={false}
-            allowClear={false}
-            onChange={handleRangePickerChange}
-          />
-        </Form.Item>
-      </div>
-    );
-  }
-}
+import InputSearch from './InputSearch';
 
 function Search(props) {
   // Set up state to keep track of search term , search criteria and modal visibility
@@ -220,7 +163,6 @@ function SearchBar(props) {
     return filteredResults;
   };
 
-  // return <Search onSubmit={handleSearchSubmit} searchResults={searchResults} setSearchResults={setSearchResults} />;
   return (
     <Search
       onSubmit={handleSearchSubmit}
